+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project will adhere to [Semantic Versioning](http://semver.org/spec/v2.
 ### Fixed
 
 * Always return predicates of list type in an array.
+* Edges without facet values are also returned when performing sort on facet.
 
 ## [1.0.1] - 2017-12-20